backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator>=2.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
boto3==1.28.62
botocore==1.31.62
urllib3==2.0.7
httpx>=0.24.0
python-socketio==5.11.4
pydantic==1.10.13